# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_lot_lots_post": {
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "date_code": {
            "title": "Date Code",
            "type": "string"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "lot_id": {
            "title": "Lot Id",
            "type": "string"
          },
          "manufacturer": {
            "title": "Manufacturer",
            "type": "string"
          },
          "part_number": {
            "title": "Part Number",
            "type": "string"
          },
          "reference_expected_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Expected Json"
          },
          "test_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Test Date"
          }
        },
        "required": [
          "file",
          "lot_id",
          "part_number",
          "manufacturer",
          "date_code",
          "account_id"
        ],
        "title": "Body_upload_lot_lots_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LoginRequest": {
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "pin": {
            "title": "Pin",
            "type": "string"
          }
        },
        "required": [
          "account_id",
          "pin"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "LotSummaryResponse": {
        "properties": {
          "assessments": {
            "items": {
              "$ref": "#/components/schemas/RiskAssessment"
            },
            "title": "Assessments",
            "type": "array"
          },
          "disposition": {
            "$ref": "#/components/schemas/LotDisposition"
          },
          "explanation_summary": {
            "default": "",
            "title": "Explanation Summary",
            "type": "string"
          },
          "insufficient_data_components": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Insufficient Data Components",
            "type": "array"
          },
          "module_a_cutoffs": {
            "additionalProperties": {
              "type": "number"
            },
            "default": {},
            "title": "Module A Cutoffs",
            "type": "object"
          },
          "module_a_results": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ModuleAResult"
            },
            "default": {},
            "title": "Module A Results",
            "type": "object"
          },
          "module_b_advisory_notes": {
            "additionalProperties": {
              "type": "string"
            },
            "default": {},
            "title": "Module B Advisory Notes",
            "type": "object"
          },
          "module_b_results": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ModuleBResult"
            },
            "default": {},
            "title": "Module B Results",
            "type": "object"
          },
          "part_explanations": {
            "additionalProperties": {
              "$ref": "#/components/schemas/PartExplanation"
            },
            "default": {},
            "title": "Part Explanations",
            "type": "object"
          }
        },
        "required": [
          "assessments",
          "disposition"
        ],
        "title": "LotSummaryResponse",
        "type": "object"
      },
      "LotUploadResponse": {
        "properties": {
          "insufficient_data_components": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Insufficient Data Components",
            "type": "array"
          },
          "lot_id": {
            "title": "Lot Id",
            "type": "string"
          },
          "part_number": {
            "title": "Part Number",
            "type": "string"
          },
          "reading_count": {
            "title": "Reading Count",
            "type": "integer"
          },
          "status": {
            "enum": [
              "IN_PROGRESS",
              "COMPLETE"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "lot_id",
          "part_number",
          "status",
          "reading_count"
        ],
        "title": "LotUploadResponse",
        "type": "object"
      },
...
      "PartDetailResponse": {
        "properties": {
          "analysis_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Run Id"
          },
          "component_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Component Id"
          },
          "confidence_qualifier": {
            "title": "Confidence Qualifier",
            "type": "string"
          },
          "confirmed_outcomes": {
            "items": {
              "$ref": "#/components/schemas/ConfirmedOutcomeRecord"
            },
            "title": "Confirmed Outcomes",
            "type": "array"
          },
          "disposition_history": {
            "items": {
              "$ref": "#/components/schemas/DispositionRecord"
            },
            "title": "Disposition History",
            "type": "array"
          },
          "disposition_status": {
            "anyOf": [
              {
                "enum": [
                  "NONE",
                  "ACCEPT_RECORDED",
                  "HOLD_RECORDED",
                  "REJECT_PENDING_SECOND",
                  "REJECT_FINAL",
                  "CONFLICT"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disposition Status"
          },
          "explanation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PartExplanation"
              },
              {
                "type": "null"
              }
            ]
          },
          "explanation_sentence": {
            "title": "Explanation Sentence",
            "type": "string"
          },
          "lot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lot Id"
          },
          "module_a": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModuleAResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "module_b": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModuleBResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "module_b_advisory_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Module B Advisory Note"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "severity_cap_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity Cap Note"
          },
          "staleness_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Staleness Note"
          },
          "unavailable_forecast_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unavailable Forecast Note"
          },
          "verdict": {
            "anyOf": [
              {
                "enum": [
                  "PASS",
                  "WATCH",
                  "REJECT"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict"
          }
        },
        "required": [
          "explanation_sentence",
          "confidence_qualifier",
          "severity_cap_note",
          "unavailable_forecast_note",
          "staleness_note",
          "disposition_history",
          "confirmed_outcomes"
        ],
        "title": "PartDetailResponse",
        "type": "object"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "token_type": {
            "const": "bearer",
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "account_id",
          "role"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Burn-In Screening API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "Auth",
          "Settings",
          "Disposition"
        ]
      }
    },
...
    "/lots": {
      "post": {
        "operationId": "upload_lot_lots_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_lot_lots_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LotUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Lot",
        "tags": [
          "ingestion"
        ]
      }
    },
...
    "/lots/{lot_id}": {
      "get": {
        "operationId": "get_lot_summary_lots__lot_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "lot_id",
            "required": true,
            "schema": {
              "title": "Lot Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LotSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Lot Summary",
        "tags": [
          "fusion"
        ]
      }
    },
...
    "/parts/{component_id}": {
      "get": {
        "operationId": "get_part_detail_parts__component_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "component_id",
            "required": true,
            "schema": {
              "title": "Component Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "lot_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Part Detail",
        "tags": [
          "fusion"
        ]
      }
    },
...
  }
}